Configure more performant cross-platform options and separate out benchmark
